pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "czarneniebo-ai"
version = "0.1.0"
description = "AI tools for investigative journalism and media verification"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Lech Rustecki", email = "czarneniebo@proton.me"}
]
keywords = ["journalism", "deepfake", "disinformation", "nlp", "polish", "ukraine"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"

# Core dependencies (public repo, Apache 2.0)
dependencies = [
    "pdfplumber>=0.10",
    "easyocr>=1.7",
    "spacy>=3.7",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "ollama>=0.3",
    "gradio>=4.0",
    "networkx>=3.1",
    "pyvis>=0.3",
    "faster-whisper>=1.0",
    "watchdog>=4.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
]

[project.optional-dependencies]
# Premium modules (private repo)
premium = [
    "facenet-pytorch>=2.5",
    "pillow>=10.0",
    "piexif>=1.1",
    "opencv-python-headless>=4.8",
    "transformers>=4.40",
    "demucs>=4.0",
]
# Docs generation
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
]
# Testing
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
# Blockchain/contracts
contracts = [
    "web3>=7.0",
    "eth-account>=0.12",
]

[project.urls]
Homepage = "https://patronite.pl/CzarneNiebo"
Repository = "https://github.com/czarneniebo/czarneniebo-ai"
Documentation = "https://czarneniebo.github.io/czarneniebo-ai"
"Bug Tracker" = "https://github.com/czarneniebo/czarneniebo-ai/issues"

[project.scripts]
czarneniebo = "czarneniebo.web_ui:main"
cn-install = "scripts.install:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["czarneniebo*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
